+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Id's must be created manually, instead of via salt.
 - Tool windows return actions which should be processed, e.g. `ToolWindowAction::CloseRequested`.
 - `Simple` demo updated to show how to use the `closable` and `titlebar_content` features.
+- Long window titles now show ellipsis if they are too long to fit.
 
 ## 0.5.0